requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
scapy>=2.5.0
psutil>=5.9.0
requests>=2.31.0
geoip2>=4.7.0
maxminddb>=2.3.0,<3.0.0
python-socketio>=5.8.0
eventlet>=0.33.0
netifaces>=0.11.0
pyyaml>=6.0